pyproject.toml
[tool.poetry]
name = "simunet"
version = "0.1.0"
readme = "README.md"
description = "mock ssh servers"
authors = ["bobby"]



[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
pyyaml = "^6.0.2"
asyncssh = "^2.17.0"
dependency-injector = "^4.42.0"
loguru = "^0.7.2"
asgi-correlation-id = "^4.3.4"
tabulate = "^0.9.0"
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
pycryptodome = "^3.21.0"
aiofiles = "^24.1.0"

[tool.poetry.scripts]
simunet = "main:start"

[tool.pytest.ini_options]
markers = ['unit', 'integration']


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# 将下方配置添加至pyproject.toml末尾即可
[[tool.poetry.source]]
name = "tsinghua-source"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true  # 仅从定义仓库获取
